fix: change column types from integer to bigint for start_at and expires_at in InitVouchers migration

+++ src/migrations/1745727699159-init-vouchers.ts
@@ -38,12 +38,12 @@ export class InitVouchers1745727699159 implements MigrationInterface {
                     },
                     {
                         name: 'start_at',
-                        type: 'integer',
+                        type: 'bigint',
                         isNullable: true,
                     },
                     {
                         name: 'expires_at',
-                        type: 'integer',
+                        type: 'bigint',
                         isNullable: true,
                     },
                     {
